Jade ver 0.3

+++ ALG/2009-jade/problem.cpp
@@ -10,24 +10,24 @@ class Problem
 {
 public:
 
-    float fun1(std::vector<float> xx, int dim){
-        float num=0;
+    long double fun1(std::vector<float> xx, int dim){
+        long double num=0;
         for (int i=1; i<=dim; i++){
             num += xx[i]*xx[i];
         }
         return num;
     }
     
-    float fun2(std::vector<float> xx, int dim){
-        float num=0;
+    long double fun2(std::vector<float> xx, int dim){
+        long double num=0;
         // todo
         
         return num;
     }
 
     
-    float fun3(std::vector<float> xx, int dim){
-        float num=0;
+    long double fun3(std::vector<float> xx, int dim){
+        long double num=0;
         // todo
         
         return num;
